+++ keras_contrib/layers/__init__.py
@@ -1,9 +1,9 @@
 from __future__ import absolute_import
 
-from .advanced_activations import PELU
-from .advanced_activations import SReLU
-from .advanced_activations import Swish
-from .advanced_activations import SineReLU
+from .advanced_activations.pelu import PELU
+from .advanced_activations.srelu import SReLU
+from .advanced_activations.swish import Swish
+from .advanced_activations.sinerelu import SineReLU
 
 from .convolutional import CosineConv2D
 from .convolutional import SubPixelUpscaling
